% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_filename_photo.R
\name{clean_filename_photo}
\alias{clean_filename_photo}
\title{Nettoyage du nom des photos}
\usage{
clean_filename_photo(nom_photo, date, nom_pe, num_uo)
}
\arguments{
\item{num_uo}{}
}
\value{

}
\description{
Nettoyage du nom des photos
}
