% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_numominidia.R
\name{clean_numominidia}
\alias{clean_numominidia}
\title{Nettoyage du code Omnidia}
\usage{
clean_numominidia(numomnidia)
}
\arguments{
\item{numomnidia}{numéro d'omnidia permet d'identifier la lame}
}
\value{
numéro d'omnidia à 9 étoiles -format chaîne de caractère-
}
\description{
Nettoyage du code Omnidia
}
\examples{
library(diatomfrlake)
clean_numominidia("170822-037")

}
\seealso{
\code{\link{formatting_numomnidia}},\code{\link{formattingdate_num}}
}
