% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_contact.R
\name{formatting_contact}
\alias{formatting_contact}
\title{Mise en forme la feuille "A renseigner" du formualire de saisie}
\usage{
formatting_contact(output_import)
}
\arguments{
\item{ods}{le fichier d'origine est un fichier avec une extension ods booléen TRUE (or FALSE)}

\item{sheet2}{sortie import_formulaire de la feuille "2"}
}
\value{

}
\description{
Mise en forme la feuille "A renseigner" du formualire de saisie
}
