% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_data.R
\name{formatting_data}
\alias{formatting_data}
\title{Mise en forme la feuille "Données générales" du formualire de saisie}
\usage{
formatting_data(output_import)
}
\arguments{
\item{output_import}{objet de sortie de la fonction \code{\link{import_formulaire}}}
}
\value{
table rempli par les données mésologiques -objet data.frame-
}
\description{
Mise en forme la feuille "Données générales" du formualire de saisie
}
