% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_flore.R
\name{formatting_flore}
\alias{formatting_flore}
\title{Mise en forme de la liste floristique dasn un forme de tableur}
\usage{
formatting_flore(flore)
}
\arguments{
\item{flore}{tableau flore sortie de read_diatom ou de la fonction read_prn}
}
\value{
suppresion de lignes identiques ; ajout de la colonne id identifiant
}
\description{
Mise en forme de la liste floristique dasn un forme de tableur
}
\examples{
library(diatomfrlake)

}
