% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_numomnidia.R
\name{formatting_numomnidia}
\alias{formatting_numomnidia}
\title{Correction des fautes de frappes lors de la saisie du numéro d'omnidia / ou - au lieu *}
\usage{
formatting_numomnidia(numomnidia)
}
\arguments{
\item{numomnidia}{numéro d'omnidia permet d'identifier la lame -format chaîne de caractère-}
}
\value{
numéro d'omnidia corrigé -format chaîne de caractère-
}
\description{
Correction des fautes de frappes lors de la saisie du numéro d'omnidia / ou - au lieu *
}
\examples{
library(diatomfrlake)
formatting_numomnidia("170822-037")

}
