% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattingdate_num.R
\name{formattingdate_num}
\alias{formattingdate_num}
\title{Formatage de la date du numéro d'omnidia à 9 étoiles aaaa-mm-dd}
\usage{
formattingdate_num(numomnidia)
}
\arguments{
\item{numomnidia}{numéro d'omnidia à 9 étoiles -format chaine de caractère-}
}
\value{
numéro d'omnidia à 9 étoiles avec la date dans le format suivant aaaa-mm-dd -format chaine de charactère-
}
\description{
Formatage de la date du numéro d'omnidia à 9 étoiles aaaa-mm-dd
}
\examples{
library(diatomfrlake)
formattingdate_num("037")
formattingdate_num("-167725522*27/07/2016********")

}
\seealso{
\code{\link{clean_numomnidia}}
}
