% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_formatOFB.R
\name{import_formatOFB}
\alias{import_formatOFB}
\title{Import data with PhytoBenthos OFB's sheet}
\usage{
import_formatOFB(file, repertoire = getwd(), sheet = 1, select = "")
}
\arguments{
\item{file}{filename of the phytobenthos OFB's sheet}

\item{repertoire}{The path of folder}
}
\value{
output
}
\description{
Import data with PhytoBenthos OFB's sheet
}
\examples{
library(diatomfrlake)

}
