% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_formulaire.R
\name{import_formulaire}
\alias{import_formulaire}
\title{Importation des données mésologiques à partir du fichier formulaire de saisie Phytobenthos Plan d'eau}
\usage{
import_formulaire(file, sheet = NA, repertoire = getwd())
}
\arguments{
\item{file}{nom du fichier à importer sous R format -chaine de charactère-}

\item{sheet}{feuille -chaine de charactère- 2 feuille A renseigner ou 3}

\item{repertoire}{dossier ou se trouve le ou les fichiers à importer}
}
\value{

}
\description{
Importation des données mésologiques à partir du fichier formulaire de saisie Phytobenthos Plan d'eau
}
\examples{
library(diatomfrlake)

}
