% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diatom.R
\name{read_diatom}
\alias{read_diatom}
\title{Importation du fichier flore (prn ou txt)}
\usage{
read_diatom(filename, repertoire = getwd(), encoding = "CP1252")
}
\arguments{
\item{filename}{nom du fichier soit .prn ou .txt}

\item{repertoire}{répertoire de données}

\item{encoding}{encoding en défaut CP1252}
}
\value{
tableau flore
}
\description{
Importation du fichier flore (prn ou txt)
}
\examples{
library(diatomfrlake)

}
