% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prn.R
\name{read_prn}
\alias{read_prn}
\title{Importation d'un fichier prn -format Omnidia-}
\usage{
read_prn(file, repertoire = getwd(), encoding = "CP1252")
}
\arguments{
\item{file}{nom de fichier Omnidia au format -prn-}

\item{repertoire}{chemin du répertoire par défaut le répertoire de travail}

\item{encoding}{code de l'encoding par défaut format windows CP1252}
}
\value{

}
\description{
Importation d'un fichier prn -format Omnidia-
}
\examples{
library(magrittr)
library(diatomfrlake)
path <- system.file("", package = "diatomfrlake")
read_prn(file ="3723.prn",repertoire=path)

}
